requirements.txt
APScheduler==3.9.1
backports.zoneinfo==0.2.1
cassandra-driver==3.25.0
certifi==2021.5.30
click==8.0.4
colorama==0.4.4
colorhash==1.0.4
configparser==5.2.0
dataclasses==0.6
Flask==2.0.3
Flask-Cors==3.0.10
Flask-MonitoringDashboard==3.0.6
geomet==0.2.1.post1
greenlet==1.1.2
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.19.5
pandas==1.1.5
psutil==5.9.0
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
scikit-learn==0.24.2
six==1.16.0
SQLAlchemy==1.4.35
threadpoolctl==3.1.0
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
Werkzeug==2.0.3
wincertstore==0.2
zipp==3.6.0
-e .